+++ R/AllGenerics.R
@@ -5,8 +5,10 @@
 ##' @name vennpie
 ##' @rdname vennpie-methods
 ##' @export
-setGeneric("vennpie", function(x, r = 0.2, cex = 1.2, ...) 
-  standardGeneric("vennpie"))
+setGeneric("vennpie", 
+  function(x, r = 0.2, cex = 1.2, ...) 
+  standardGeneric("vennpie")
+)
 
 
 ##' plotDistToTSS method generics
@@ -16,10 +18,14 @@ setGeneric("vennpie", function(x, r = 0.2, cex = 1.2, ...)
 ##' @name plotDistToTSS
 ##' @rdname plotDistToTSS-methods
 ##' @export
-setGeneric("plotDistToTSS", function(x, distanceColumn="distanceToTSS",
-                                     xlab="", ylab="Binding sites (%) (5'->3')",
-                                     title="Distribution of transcription factor-binding loci relative to TSS", ...)
-           standardGeneric("plotDistToTSS"))
+setGeneric("plotDistToTSS", 
+  function(x, 
+    distanceColumn="distanceToTSS",
+    xlab="", ylab="Binding sites (%) (5'->3')",
+    title="Distribution of transcription factor-binding loci relative to TSS", 
+    ...)
+  standardGeneric("plotDistToTSS")
+)
 
 ##' plotAnnoBar method generics
 ##'
@@ -28,12 +34,14 @@ setGeneric("plotDistToTSS", function(x, distanceColumn="distanceToTSS",
 ##' @name plotAnnoBar
 ##' @rdname plotAnnoBar-methods
 ##' @export
-setGeneric("plotAnnoBar", function(x,
-                                   xlab="",
-                                   ylab="Percentage(%)",
-                                   title="Feature Distribution",
-                                   ...)
-           standardGeneric("plotAnnoBar"))
+setGeneric("plotAnnoBar", 
+  function(x,
+    xlab="",
+    ylab="Percentage(%)",
+    title="Feature Distribution",
+    ...)
+  standardGeneric("plotAnnoBar")
+)
 
 
 ##' plotAnnoPie method generics
@@ -44,12 +52,13 @@ setGeneric("plotAnnoBar", function(x,
 ##' @rdname plotAnnoPie-methods
 ##' @export
 setGeneric("plotAnnoPie", 
-           function(x, 
-                    ndigit=2,
-                    cex=0.9,
-                    col=NA,
-                    legend.position="rightside",
-                    pie3D=FALSE,
-                    radius=0.8,
-                    ...)
-           standardGeneric("plotAnnoPie"))
+  function(x, 
+    ndigit=2,
+    cex=0.9,
+    col=NA,
+    legend.position="rightside",
+    pie3D=FALSE,
+    radius=0.8,
+    ...)
+  standardGeneric("plotAnnoPie")
+)