+++ R/covplot.R
@@ -11,7 +11,7 @@
 ##' @param chrs selected chromosomes to plot, all chromosomes by default
 ##' @param xlim ranges to plot, default is whole chromosome
 ##' @param lower lower cutoff of coverage signal
-##' @param fill_color specify the color for the plot. Order matters
+##' @param fill_color specify the color/palette for the plot. Order matters
 ##' @return ggplot2 object
 ##' @import GenomeInfoDb
 ##' @importFrom ggplot2 ggplot
@@ -35,35 +35,20 @@ covplot <- function(peak, weightCol=NULL,
                     chrs  = NULL,
                     xlim  = NULL,
                     lower = 1,
-                    fill_color = NULL) {
-    
-    isList <- FALSE
-    if(is(peak, "GRanges") || length(peak) == 1) {
-        tm <- getChrCov(peak=peak, weightCol=weightCol, chrs, xlim, lower=lower)
-        
-        if(is.null(fill_color)){
-            fill_color = 'black'
-        }else{
-            if(length(fill_color) != 1){
-                stop('pls input fill_color parameter with correct length...')
-            }
-        }
-        
+                    fill_color = "black") {
+    isList <- is.list(peak)
+    if(!isList) {  # Note: don't support data.frame
+        tm <- getChrCov(peak = peak, weightCol = weightCol, chrs = chrs, xlim = xlim, lower = lower)
     } else {
-        isList <- TRUE
-        ltm <- lapply(peak, getChrCov, weightCol=weightCol, chrs=chrs, xlim=xlim, lower=lower)
+        ltm <- lapply(peak, getChrCov, weightCol = weightCol, chrs = chrs, xlim = xlim, lower = lower)
         if (is.null(names(ltm))) {
             nn <- paste0("peak", seq_along(ltm))
-            warning("input is not a named list, set the name automatically to ", paste(nn, collapse=' '))
+            warning("input is not a named list, set the name automatically to ", paste(nn, collapse = ' '))
             names(ltm) <- nn
         }
-        tm <- list_to_dataframe(ltm)
+        tm <- dplyr::bind_rows(ltm, .id = ".id")
         chr.sorted <- sortChrName(as.character(unique(tm$chr)))
         tm$chr <- factor(tm$chr, levels = chr.sorted)
-        
-        if(!is.null(fill_color) && length(fill_color) != length(unique(tm$.id))){
-            stop('pls input fill_color parameter with correct length...')
-        }
     }
     
     chr <- start <- end <- value <- .id <- NULL
@@ -75,14 +60,16 @@ covplot <- function(peak, weightCol=NULL,
         
         ## p <- p + geom_segment(aes(x=start, y=0, xend=end, yend= value))
         if (isList) {
-            
-            p <- p + geom_rect(aes(xmin=start, ymin=0, xmax=end, ymax=value, fill=.id, color=.id)) 
-            if(!is.null(fill_color)){
-                p <- p + scale_fill_manual(values = fill_color) + scale_color_manual(values = fill_color)
+            if (length(fill_color) == length(peak) && all(is_valid_color(fill_color))){
+                cols = fill_color
+            } else {
+                cols = generate_colors(fill_color, n = length(peak))
             }
-            
+            p <- p + geom_rect(aes(xmin = start, ymin = 0, xmax = end, ymax = value, fill = .id, color = .id)) +
+                scale_color_manual(values = cols) +
+                scale_fill_manual(values = cols)
         } else {
-            p <- p + geom_rect(aes(xmin=start, ymin=0, xmax=end, ymax=value), fill=fill_color, color=fill_color)
+            p <- p + geom_rect(aes(xmin = start, ymin = 0, xmax = end, ymax = value), fill = fill_color, color = fill_color)
         }
         
         if(length(unique(tm$chr)) > 1) {
@@ -92,9 +79,10 @@ covplot <- function(peak, weightCol=NULL,
     }
     
     p <- p + theme_classic()
-    p <- p + xlab(xlab) + ylab(ylab) + ggtitle(title)
-    p <- p + scale_y_continuous(expand=c(0,0))
+    p <- p + labs(x = xlab, y = ylab, title = title, fill = NULL, color = NULL)
+    p <- p + scale_y_continuous(expand = c(0,0))
     p <- p + theme(strip.text.y=element_text(angle=360))
+    p <- p + scale_x_continuous(labels = scales::label_number(scale_cut = scales::cut_si("")))
     
     if (!is.null(xlim) && !all(is.na(xlim)) && is.numeric(xlim) && length(xlim) == 2) {
         p <- p + xlim(xlim)
@@ -123,7 +111,7 @@ getChrCov <- function(peak, weightCol, chrs, xlim, lower=1) {
         peak.cov <- coverage(peak.gr, weight=weight)
     }
 
-    cov <- lapply(peak.cov, slice, lower=lower)
+    cov <- lapply(peak.cov, IRanges::slice, lower=lower)
 
     get.runValue <- function(x) {
         y <- runValue(x)
@@ -165,23 +153,16 @@ getChrCov <- function(peak, weightCol, chrs, xlim, lower=1) {
         df <- df[df$start >= xlim[1] & df$end <= xlim[2],]
     }
 
-    df2 <- group_by(df, chr, start, end) %>% summarise(value=sum(cnt))
+    df2 <- group_by(df, chr, start, end) %>% summarise(value=sum(cnt), .groups = "drop")
     return(df2)
 }
 
-sortChrName <- function(chr.name) {
-    ## X, Y and M will cause warnings when change to number.
-    noChr <- suppressWarnings(as.numeric(sub("chr", "", chr.name)))
-    ## index of chromosome name are character, such as X, Y
-    ch.idx <- which(is.na(noChr))
-    
-    n.idx <- which(!is.na(noChr))
-    chr.n <- noChr[n.idx]
-
-    chr.sorted <- chr.name[n.idx][order(chr.n)]
-    if (length(ch.idx) != 0) {
-        chr.sorted <- c(chr.sorted, sort(chr.name[ch.idx]))
-    }
-         
-    return(chr.sorted)
+# a simple `stringr::str_sort(numeric=TRUE)` implementation
+sortChrName <- function(chr.name, decreasing = FALSE) {
+    ## universal sort function, support organisms other than human
+    chr_part <- sub("^(\\D*)(\\d*)$", "\\1", chr.name)
+    num_part <- as.numeric(sub("^(\\D*)(\\d*)$", "\\2", chr.name))
+    chr.name[order(chr_part, num_part, decreasing = decreasing)]
 }
+
+